# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUS AI - RPG-style AI skill system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all acquired skills
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Manually create a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get skill detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404":
          description: Not found
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /skills/study:
    post:
      operationId: studyTopic
      tags: [skills]
      summary: Ask AI to study a topic or repository and propose skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyRequest"
      responses:
        "200":
          description: Proposed skills from study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyResult"
...
  /skills/fuse:
    post:
      operationId: fuseSkills
      tags: [skills]
      summary: Fuse two or more skills into a new combined skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FuseRequest"
      responses:
        "200":
          description: Fused skill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuseResult"
...
components:
  schemas:
...
    Skill:
      type: object
      required: [id, name, category, description, level, status, xpValue, icon, color, principles, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        level:
          type: integer
        status:
          type: string
          description: "acquired | pending | fused"
        xpValue:
          type: integer
        icon:
          type: string
        color:
          type: string
        principles:
          type: array
          items:
            type: string
        parentSkillIds:
          type: ["array", "null"]
          items:
            type: integer
        studySource:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SkillInput:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
    SkillUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
    StudyRequest:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          description: Topic, URL, or repository to study
        filter:
          type: ["string", "null"]
          description: What to include/exclude from study
    StudyResult:
      type: object
      required: [summary, proposedSkills]
      properties:
        summary:
          type: string
        proposedSkills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
    FuseRequest:
      type: object
      required: [skillIds]
      properties:
        skillIds:
          type: array
          items:
            type: integer
          minItems: 2
        fusionName:
          type: ["string", "null"]
    FuseResult:
      type: object
      required: [fusedSkill, description]
      properties:
        fusedSkill:
          $ref: "#/components/schemas/Skill"
        description:
          type: string
...